+++ libs/components/toast/projects/toast/src/public-api.ts
@@ -5,3 +5,7 @@ export * from './modules/toast/types/toast-config';
 export * from './modules/toast/types/toast-type';
 export * from './modules/toast/types/toast-container-options';
 export * from './modules/toast/types/toast-display-direction';
+
+// Components and directives must be exported to support Angular's "partial" Ivy compiler.
+// Obscure names are used to indicate types are not part of public API.
+export { SkyToastComponent as λ1 } from './modules/toast/toast.component';