+++ libs/components/toast/CHANGELOG.md
@@ -1,3 +1,21 @@
+# 5.0.0-beta.2 (2021-09-10)
+
+- Enabled Ivy's "partial" compilation mode. [#73](https://github.com/blackbaud/skyux-toast/pull/73)
+
+# 5.0.0-beta.1 (2021-08-30)
+
+- Migrated to Angular CLI. [#70](https://github.com/blackbaud/skyux-toast/pull/70)
+
+# 5.0.0-beta.0 (2021-06-29)
+
+- Initial beta release.
+- Added support for `5.0.0-beta.*` versions of SKY UX component libraries. [#68](https://github.com/blackbaud/skyux-toast/pull/68)
+- Fixed `SkyToastService` to work within lazy-loaded feature modules. [#68](https://github.com/blackbaud/skyux-toast/pull/68)
+
+# 5.0.0-alpha.0 (2021-05-21)
+
+- Removed `BrowserAnimationsModule` from the `imports` section of `SkyToastModule` to support lazy-loading. Consumers of `SkyToastModule` must now import `BrowserAnimationsModule` into their application's root module. [#67](https://github.com/blackbaud/skyux-toast/pull/67)
+
 # 4.1.2 (2020-11-23)
 
 - Updated the tooltip for the button to close toasts. [#64](https://github.com/blackbaud/skyux-toast/pull/64)